atm simulation - stage 1

+++ src/entity/Account.java
@@ -3,13 +3,13 @@ package entity;
 public class Account {
     private String name;
     private String pin;
-    private double balance;
+    private int balance;
     private String accountNumber;
 
     public Account() {
     }
 
-    public Account(String name, String pin, double balance, String accountNumber) {
+    public Account(String name, String pin, int balance, String accountNumber) {
         this.name = name;
         this.pin = pin;
         this.balance = balance;
@@ -32,11 +32,11 @@ public class Account {
         this.pin = pin;
     }
 
-    public double getBalance() {
+    public int getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(int balance) {
         this.balance = balance;
     }
